Fixes audio issues when using CMake on Mac
