+++ cocos/CMakeLists.txt
@@ -110,6 +110,8 @@ set(COCOS2D_BASE_SRC
     base/pvr.cpp
     base/s3tc.cpp
 
+    audio/AudioEngine.cpp
+
     2d/CCAction.cpp
     2d/CCActionCamera.cpp
     2d/CCActionCatmullRom.cpp
@@ -427,6 +429,9 @@ elseif(MACOSX)
         ${COCOS2D_AUDIO_SRC_C}
         audio/mac/SimpleAudioEngine.mm
         audio/mac/CDXMacOSXSupport.mm
+        audio/apple/AudioCache.mm
+        audio/apple/AudioEngine-inl.mm
+        audio/apple/AudioPlayer.mm
        )
     SET_SOURCE_FILES_PROPERTIES(
         ${COCOS2D_AUDIO_SRC_C}
